requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0  # Async PostgreSQL driver
alembic==1.12.1
psycopg2-binary==2.9.9  # For Alembic sync operations

# Environment & Security
python-dotenv==1.0.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# Data Processing & Export
pandas==2.1.4
openpyxl==3.1.2
weasyprint==60.1

# HTMX & Templates
jinja2==3.1.2
